Update the screen reducer
The screen reducer now keeps track of shapes in two ways.

Saved Shapes: Shapes that are drawn and the user selects ‘close and save’

Actions: This is a list of drawing actions the user has done while in edit mode. Basically every time the user draws, deletes or edits a shape, that action gets recorded in state. When displaying shapes, we apply all of the actions on top of the saved shapes. When the user saves their work, the saved shapes gets updated with all of drawing actions